Cleanup: Remove helper Python scripts used during refactor
- Removed temporary scripts created to automate package and import migrations.
