+++ app/src/androidTest/java/org/maocide/undeadwallpaper/data/VideoFileManagerBenchmarkTest.kt
@@ -1,4 +1,6 @@
-package org.maocide.undeadwallpaper
+package org.maocide.undeadwallpaper.data
+
+import org.maocide.undeadwallpaper.R
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
