+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
         android:theme="@style/Theme.UndeadWallpaper"
         tools:targetApi="31">
         <service
-            android:name="UndeadWallpaperService"
+            android:name=".service.UndeadWallpaperService"
             android:thumbnail="@mipmap/ic_launcher" android:description="@string/app_name"
             android:exported="true"
             android:enabled="true"
@@ -49,7 +49,7 @@
         </service>
 
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.UndeadWallpaper">
             <intent-filter>
