+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="org.maocide.undeadwallpaper.SettingsFragment"
+        android:name="org.maocide.undeadwallpaper.ui.SettingsFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_settings">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="org.maocide.undeadwallpaper.AboutFragment"
+        android:name="org.maocide.undeadwallpaper.ui.AboutFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_about">
