+++ app/src/main/res/xml/wallpaper.xml
@@ -3,4 +3,4 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:thumbnail="@drawable/wallpaper_preview"
     android:description="@string/wallpaper_description"
-    android:settingsActivity="org.maocide.undeadwallpaper.MainActivity"/>
+    android:settingsActivity="org.maocide.undeadwallpaper.ui.MainActivity"/>
